requirements.txt
boto3
requests
beautifulsoup4
python-dotenv
selenium
flask
flask-cors 